Improve readability by separing concerns
